Moved reservations out of the dashboard, into its own component
