+++ front-end/src/dashboard/Dashboard.js
@@ -4,6 +4,7 @@ import ErrorAlert from "../layout/ErrorAlert";
 import { today, previous, next } from "../utils/date-time";
 import { useLocation, useHistory } from "react-router-dom/cjs/react-router-dom";
 import ListAllTables from "./ListAllTables";
+import ListAllReservations from "./ListAllReservations";
 
 /**
  * Defines the dashboard page.
@@ -46,28 +47,9 @@ function Dashboard({ date }) {
         break;
       default:
         history.push(`/dashboard/?date=${today()}`);
-        // setCurrDate(today());
         break;
     }
   }
-
-  const tableRows = reservations.map((reservation) => {
-    return (
-      <tr key={reservation.reservation_id}>
-        <td>{reservation.first_name}</td>
-        <td>{reservation.last_name}</td>
-        <td>{reservation.mobile_number}</td>
-        <td>{reservation.reservation_date}</td>
-        <td>{reservation.reservation_time}</td>
-        <td>{reservation.people}</td>
-        <td>
-          <a href={`/reservations/${reservation.reservation_id}/seat`}>
-            <button>Seat</button>
-          </a>
-        </td>
-      </tr>
-    );
-  });
 
   return (
     <main>
@@ -78,20 +60,8 @@ function Dashboard({ date }) {
           <h4 className="mb-0">Reservations for {currDate}</h4>
         </div>
         <ErrorAlert error={reservationsError} />
-        <table className="table">
-          <thead>
-            <tr>
-              <th>First Name</th>
-              <th>Last Name</th>
-              <th>Mobile Number</th>
-              <th>Reservation Date</th>
-              <th>Reservation Time</th>
-              <th>Party Size</th>
-              <th>Seat</th>
-            </tr>
-          </thead>
-          <tbody>{tableRows}</tbody>
-        </table>
+
+        <ListAllReservations reservations={reservations} />
         <div>
           <button onClick={buttonHandler} name="previous">
             Previous
